requirement.txt
Flask==3.1.0
flask-cors==5.0.1
Flask-Session==0.8.0
requests==2.32.3
bcrypt==4.3.0
scikit-learn==1.6.1
mysql-connector-python==9.3.0
python-dotenv==1.0.1